Refactor so that TopMainNav components are exported more pleasantly
