+++ src/components/Header/TopMainNav/Dropdown/Button/AccountPanelDropdownData.tsx
@@ -1,4 +1,4 @@
-import { AccountPanel } from './AccountPanel';
+import { AccountPanel } from './AccountPanel/index';
 import { DropdownData } from '../types';
 
 export const AccountPanelDropdownData: DropdownData = {
