+++ src/components/Header/TopMainNav/Dropdown/Summary/index.ts
@@ -1,0 +1,2 @@
+export * from './Summary';
+export * from './SummaryLinkIcon';
