+++ src/components/Header/TopMainNav/Dropdown/index.ts
@@ -1,0 +1,3 @@
+export * from './types';
+export * from './DropdownMenu';
+export * from './DropdownMenuContainer';
