+++ src/components/Header/TopMainNav/SearchBar/index.ts
@@ -1,0 +1,2 @@
+export * from './SearchBar';
+export * from './SearchIcon';
