+++ src/components/Header/TopMainNav/TopMainNavUser/SignedIn/index.ts
@@ -1,0 +1,3 @@
+export * from './SignedInPlaceholder';
+export * from './SignIn';
+export * from './SignOutLink';
