+++ src/components/Header/TopMainNav/index.ts
@@ -1,0 +1,1 @@
+export * from './TopMainNav';
